+++ sdk/trace/batch_span_processor_test.go
@@ -23,6 +23,7 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/otel/trace"
 
@@ -287,6 +288,31 @@ func TestBatchSpanProcessorShutdown(t *testing.T) {
 		t.Error("Error shutting the BatchSpanProcessor down\n")
 	}
 	assert.Equal(t, 1, bp.shutdownCount)
+}
+
+func TestBatchSpanProcessorPostShutdown(t *testing.T) {
+	tp := basicTracerProvider(t)
+	be := testBatchExporter{}
+	bsp := sdktrace.NewBatchSpanProcessor(&be)
+
+	tp.RegisterSpanProcessor(bsp)
+	tr := tp.Tracer("Normal")
+
+	generateSpan(t, true, tr, testOption{
+		o: []sdktrace.BatchSpanProcessorOption{
+			sdktrace.WithMaxExportBatchSize(50),
+		},
+		genNumSpans: 60,
+	})
+
+	require.NoError(t, bsp.Shutdown(context.Background()), "shutting down BatchSpanProcessor")
+	lenJustAfterShutdown := be.len()
+
+	_, span := tr.Start(context.Background(), "foo")
+	span.End()
+	assert.NoError(t, bsp.ForceFlush(context.Background()), "force flushing BatchSpanProcessor")
+
+	assert.Equal(t, lenJustAfterShutdown, be.len(), "OnEnd and ForceFlush should have no effect after Shutdown")
 }
 
 func TestBatchSpanProcessorForceFlushSucceeds(t *testing.T) {